Rename AssertSameCount to AssertCompare
