+++ tests/Rector/Contrib/PHPUnit/SpecificMethod/AssertCompareToSpecificMethodRector/Correct/correct.php.inc
